Login/Password page
worked on the Authview to include a log in for user tied to our firebase for auth. Created a registerView which holds the logic for user creation and saving user data to db. worked on forgot password page where the user can enter their email and temporary password is set, then sent to user for log in. also added the assets folder from our previous project so we have a logo. also worked in authview model to define data inputed by the user to be registered. refer to fcomments in code to see what it does.