Created StatsView placeholder and created navigation bar
The Sidemenu has now become a navigation bar that is shown all the time. Signout has been moved to the ProfileView screen and has a confirmation message. I removed the breakpoints in the code as it was causing problems with the logout feature. Stats page has been created with a placeholder.