Final QA submission
worked on overall app design and quality. added a user card to the dashboard, redesigned the calendar, added a red accent to the whole app. fixed user profile logic to fetch username and edit. same for weight and height. conversion functions were created to allow for use of lbs and inches. redesigned workout page allow for better viewing of workouts. redesigned states page to include a streak feature. picture is now functional.